refactoring file locations to cleanup
